smartEQ: Add custom CAN write command and door lock metrics
Introduced the xsq canwrite command for sending custom CAN messages, with documentation and command handler. Added support for reading and tracking door lock status from BCM EEPROM. Refactored command implementations into a new eq_cmds_override.cpp file and updated CMakeLists.txt accordingly. Improved documentation for DDT4all and CAN write usage.
